Festin Joyeaux 1732 - first fifteen tunes
The first fifteen of 49 tunes in the Festin Joyeaux collection of 1732.
The book is a collection of menus for large banquets.  As courses are served, the dishes are presumably brought out as a staff member sings the list of dishes being served.  Tune #3 has just one stanza of words set to the music as an example.  (That particular course covered several stanzas!)